Change horarioLen type from uint32_t to uint8_t; update parsing to handle float values for hourStart and horarioLen

+++ main/inc/typedef.h
@@ -171,7 +171,7 @@ PV1V
 
 typedef struct {
     uint8_t hourStart;
-    uint32_t horarioLen;
+    uint8_t horarioLen;
     uint8_t pwmDuty;
 } horario_t;
